+++ jest.globalSetup.js
@@ -1,3 +1,0 @@
-module.exports = () => {
-  process.env.JUPITERONE_RUNTIME_ENVIRONMENT = "LOCAL";
-};