+++ src/crowdstrike/FalconAPIClient.test.ts
@@ -1,23 +1,18 @@
-import { IntegrationLogger } from "@jupiterone/jupiter-managed-integration-sdk";
-import { Polly } from "@pollyjs/core";
+import { IntegrationLogger } from '@jupiterone/integration-sdk-core';
+import { createMockIntegrationLogger } from '@jupiterone/integration-sdk-testing';
 
-import polly from "../../test/helpers/polly";
-import config from "../../test/integrationInstanceConfig";
-import { FalconAPIClient } from "./FalconAPIClient";
+import {
+  Recording,
+  setupCrowdstrikeRecording,
+} from '../../test/helpers/recording';
+import config from '../../test/integrationInstanceConfig';
+import { FalconAPIClient } from './FalconAPIClient';
 
 function createTestLogger(): IntegrationLogger {
-  return {
-    trace: () => undefined,
-    debug: () => undefined,
-    info: () => undefined,
-    warn: () => undefined,
-    error: () => undefined,
-    fatal: () => undefined,
-    child: () => createTestLogger(),
-  };
+  return createMockIntegrationLogger();
 }
 
-let p: Polly;
+let recording: Recording;
 
 const createClient = (): FalconAPIClient => {
   return new FalconAPIClient({
@@ -27,15 +22,20 @@ const createClient = (): FalconAPIClient => {
 };
 
 afterEach(async () => {
-  await p.stop();
+  if (recording) {
+    await recording.stop();
+  }
 });
 
-describe("authenticate", () => {
-  test("obtains token and expiration", async () => {
-    p = polly(__dirname, "authenticate");
+describe('authenticate', () => {
+  test('obtains token and expiration', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'authenticate',
+    });
 
     let requests = 0;
-    p.server.any().on("request", (_req, _event) => {
+    recording.server.any().on('request', (_req, _event) => {
       requests++;
     });
 
@@ -47,11 +47,14 @@ describe("authenticate", () => {
     expect(requests).toEqual(1);
   });
 
-  test("answers cached token before expiration", async () => {
-    p = polly(__dirname, "authenticateCached");
+  test('answers cached token before expiration', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'authenticateCached',
+    });
 
     let requests = 0;
-    p.server.any().on("request", (_req, _event) => {
+    recording.server.any().on('request', (_req, _event) => {
       requests++;
     });
 
@@ -62,11 +65,14 @@ describe("authenticate", () => {
     expect(requests).toEqual(1);
   });
 
-  test("answers new token after expiration", async () => {
-    p = polly(__dirname, "authenticateRefresh");
+  test('answers new token after expiration', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'authenticateRefresh',
+    });
 
     let requests = 0;
-    p.server.any().on("request", (_req, _event) => {
+    recording.server.any().on('request', (_req, _event) => {
       requests++;
     });
 
@@ -75,7 +81,7 @@ describe("authenticate", () => {
 
     const realNow = Date.now; // eslint-disable-line @typescript-eslint/unbound-method
     jest
-      .spyOn(global.Date, "now")
+      .spyOn(global.Date, 'now')
       .mockImplementationOnce(() => realNow() + 30 * 60 * 1000);
 
     const newAccess = await client.authenticate();
@@ -88,12 +94,19 @@ describe("authenticate", () => {
     expect(requests).toEqual(2);
   });
 
-  test("throws errors", async () => {
-    p = polly(__dirname, "authenticateError", { recordFailedRequests: true });
+  test('throws errors', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'authenticateError',
+      options: {
+        recordFailedRequests: true,
+      },
+    });
+
     const client = new FalconAPIClient({
       credentials: {
         ...config,
-        clientSecret: "test-error-handling",
+        clientSecret: 'test-error-handling',
       },
       logger: createTestLogger(),
     });
@@ -107,16 +120,19 @@ describe("authenticate", () => {
   });
 });
 
-describe("executeAPIRequest", () => {
-  test("waits until retryafter on 500 response", async () => {
-    p = polly(__dirname, "executeAPIRequest429");
+describe('executeAPIRequest', () => {
+  test('waits until retryafter on 500 response', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'executeAPIRequest429',
+    });
 
     const requestTimes: number[] = [];
-    p.server.any().on("request", (_req, _event) => {
+    recording.server.any().on('request', (_req, _event) => {
       requestTimes.push(Date.now());
     });
 
-    p.server
+    recording.server
       .any()
       .times(1)
       .intercept((_req, res) => {
@@ -128,29 +144,32 @@ describe("executeAPIRequest", () => {
     expect(requestTimes.length).toBe(2);
   });
 
-  test("waits until retryafter on 429 response", async () => {
-    p = polly(__dirname, "executeAPIRequest429");
+  test('waits until retryafter on 429 response', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'executeAPIRequest429',
+    });
 
     const requestTimes: number[] = [];
-    p.server.any().on("request", (_req, _event) => {
+    recording.server.any().on('request', (_req, _event) => {
       requestTimes.push(Date.now());
     });
 
     const retryAfter = Date.now() + 1000;
-    p.server
+    recording.server
       .any()
       .times(1)
       .intercept((_req, res) => {
         res
           .status(429)
           .setHeaders({
-            "x-ratelimit-retryafter": String(retryAfter),
+            'x-ratelimit-retryafter': String(retryAfter),
           })
           .json({
             errors: [
               {
                 code: 429,
-                message: "API rate limit exceeded.",
+                message: 'API rate limit exceeded.',
               },
             ],
           });
@@ -162,25 +181,28 @@ describe("executeAPIRequest", () => {
     expect(requestTimes[1]).toBeGreaterThan(retryAfter);
   });
 
-  test("retries 429 response limited times", async () => {
-    p = polly(__dirname, "executeAPIRequest429limit");
+  test('retries 429 response limited times', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'executeAPIRequest429limit',
+    });
 
     const requestTimes: number[] = [];
-    p.server.any().on("request", (_req, _event) => {
+    recording.server.any().on('request', (_req, _event) => {
       requestTimes.push(Date.now());
     });
 
-    p.server.any().intercept((_req, res) => {
+    recording.server.any().intercept((_req, res) => {
       res
         .status(429)
         .setHeaders({
-          "x-ratelimit-retryafter": String(Date.now() - 10),
+          'x-ratelimit-retryafter': String(Date.now() - 10),
         })
         .json({
           errors: [
             {
               code: 429,
-              message: "API rate limit exceeded.",
+              message: 'API rate limit exceeded.',
             },
           ],
         });
@@ -199,24 +221,27 @@ describe("executeAPIRequest", () => {
     expect(requestTimes.length).toBe(2);
   });
 
-  test("throttles at specified reserveLimit", async () => {
-    p = polly(__dirname, "executeAPIRequestReserveLimit");
+  test('throttles at specified reserveLimit', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'executeAPIRequestReserveLimit',
+    });
 
     let limitRemaining = 10;
 
-    p.server.any().intercept((_req, res) => {
+    recording.server.any().intercept((_req, res) => {
       limitRemaining--;
       res
         .status(201)
         .setHeaders({
-          "x-ratelimit-limit": "10",
-          "x-ratelimit-remaining": String(limitRemaining),
+          'x-ratelimit-limit': '10',
+          'x-ratelimit-remaining': String(limitRemaining),
         })
         .json({
           errors: [
             {
               code: 429,
-              message: "API rate limit exceeded.",
+              message: 'API rate limit exceeded.',
             },
           ],
         });
@@ -241,9 +266,13 @@ describe("executeAPIRequest", () => {
   });
 });
 
-describe("iterateDevices", () => {
-  test("complete set in single callback", async () => {
-    p = polly(__dirname, "iterateDevicesSinglePage");
+describe('iterateDevices', () => {
+  test('complete set in single callback', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iterateDevicesSinglePage',
+    });
+
     const cbSpy = jest.fn();
 
     const paginationState = await createClient().iterateDevices({
@@ -267,8 +296,12 @@ describe("iterateDevices", () => {
     );
   }, 20000);
 
-  test("partial set in multiple callbacks", async () => {
-    p = polly(__dirname, "iterateDevicesCompletes");
+  test('partial set in multiple callbacks', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iterateDevicesCompletes',
+    });
+
     const cbSpy = jest.fn();
 
     const paginationState = await createClient().iterateDevices({
@@ -297,8 +330,12 @@ describe("iterateDevices", () => {
     ]);
   }, 20000);
 
-  test("partial set", async () => {
-    p = polly(__dirname, "iterateDevicesInterrupted");
+  test('partial set', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iterateDevicesInterrupted',
+    });
+
     const cbSpy = jest.fn().mockResolvedValue(false);
 
     const paginationState = await createClient().iterateDevices({
@@ -322,8 +359,12 @@ describe("iterateDevices", () => {
     ]);
   }, 20000);
 
-  test("resumes from pagination state", async () => {
-    p = polly(__dirname, "iterateDevicesResumes");
+  test('resumes from pagination state', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iterateDevicesResumes',
+    });
+
     const client = createClient();
     const cbSpy = jest.fn().mockResolvedValueOnce(false);
     const paginationState = await client.iterateDevices({
@@ -353,8 +394,12 @@ describe("iterateDevices", () => {
     expect(cbSpy).toHaveBeenCalledTimes(3);
   }, 20000);
 
-  test("pagination with filter", async () => {
-    p = polly(__dirname, "iterateDevicesFilter");
+  test('pagination with filter', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iterateDevicesFilter',
+    });
+
     const client = createClient();
     const cbSpy = jest.fn().mockResolvedValueOnce(false);
 
@@ -400,10 +445,15 @@ describe("iterateDevices", () => {
     ]);
   }, 20000);
 
-  test("throws error on expired pagination offset cursor", async () => {
-    p = polly(__dirname, "iterateDevicesExpiredOffsetCursor", {
-      recordFailedRequests: true,
+  test('throws error on expired pagination offset cursor', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iterateDevicesExpiredOffsetCursor',
+      options: {
+        recordFailedRequests: true,
+      },
     });
+
     const cbSpy = jest.fn();
     await expect(
       createClient().iterateDevices({
@@ -411,7 +461,7 @@ describe("iterateDevices", () => {
         pagination: {
           limit: 1,
           offset:
-            "DnF1ZXJ5VGhlbkZldGNoBQAAAAA6iW-VFnpjQldDMjNiU2htV1FGbUo5anc1d0EAAAAAOnTgtRZqV0xVRF9ndFFhQ09zYkVhTzNNOUxnAAAAADpz2L8WSWdUN2d1OVBUM2kxNGNVMUlaU1BZUQAAAAA6hrG3FlR6aHpFU1UxUkNpZ3FVV2tWNVBkQ2cAAAAAOnto1hZTajRqVll0dVF3U2VQQThXdlo3ZjZB",
+            'DnF1ZXJ5VGhlbkZldGNoBQAAAAA6iW-VFnpjQldDMjNiU2htV1FGbUo5anc1d0EAAAAAOnTgtRZqV0xVRF9ndFFhQ09zYkVhTzNNOUxnAAAAADpz2L8WSWdUN2d1OVBUM2kxNGNVMUlaU1BZUQAAAAA6hrG3FlR6aHpFU1UxUkNpZ3FVV2tWNVBkQ2cAAAAAOnto1hZTajRqVll0dVF3U2VQQThXdlo3ZjZB',
           expiresAt: Date.now() - 1,
         },
       }),
@@ -419,9 +469,13 @@ describe("iterateDevices", () => {
   });
 });
 
-describe("iteratePreventionPolicies", () => {
-  test("complete set in single callback", async () => {
-    p = polly(__dirname, "iteratePreventionPoliciesSinglePage");
+describe('iteratePreventionPolicies', () => {
+  test('complete set in single callback', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iteratePreventionPoliciesSinglePage',
+    });
+
     const cbSpy = jest.fn();
 
     const paginationState = await createClient().iteratePreventionPolicies({
@@ -448,8 +502,12 @@ describe("iteratePreventionPolicies", () => {
     );
   }, 20000);
 
-  test("partial set in multiple callbacks", async () => {
-    p = polly(__dirname, "iteratePreventionPoliciesCompletes");
+  test('partial set in multiple callbacks', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iteratePreventionPoliciesCompletes',
+    });
+
     const cbSpy = jest.fn();
 
     const paginationState = await createClient().iteratePreventionPolicies({
@@ -487,8 +545,12 @@ describe("iteratePreventionPolicies", () => {
     ]);
   }, 20000);
 
-  test("partial set", async () => {
-    p = polly(__dirname, "iteratePreventionPoliciesInterrupted");
+  test('partial set', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iteratePreventionPoliciesInterrupted',
+    });
+
     const cbSpy = jest.fn().mockResolvedValue(false);
 
     const paginationState = await createClient().iteratePreventionPolicies({
@@ -511,8 +573,12 @@ describe("iteratePreventionPolicies", () => {
     ]);
   }, 20000);
 
-  test("resumes from pagination state", async () => {
-    p = polly(__dirname, "iteratePreventionPoliciesResumes");
+  test('resumes from pagination state', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iteratePreventionPoliciesResumes',
+    });
+
     const client = createClient();
     const cbSpy = jest.fn().mockResolvedValueOnce(false);
     const paginationState = await client.iteratePreventionPolicies({
@@ -541,8 +607,12 @@ describe("iteratePreventionPolicies", () => {
     expect(cbSpy).toHaveBeenCalledTimes(6);
   }, 20000);
 
-  test("pagination with filter", async () => {
-    p = polly(__dirname, "iteratePreventionPoliciesFilter");
+  test('pagination with filter', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iteratePreventionPoliciesFilter',
+    });
+
     const client = createClient();
     const cbSpy = jest.fn().mockResolvedValueOnce(false);
 
@@ -588,17 +658,20 @@ describe("iteratePreventionPolicies", () => {
   }, 20000);
 });
 
-describe("iteratePreventionPolicyMemberIds", () => {
-  test("complete set in single callback", async () => {
-    p = polly(__dirname, "iteratePreventionPolicyMemberIdsSinglePage");
+describe('iteratePreventionPolicyMemberIds', () => {
+  test('complete set in single callback', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iteratePreventionPolicyMemberIdsSinglePage',
+    });
+
     const cbSpy = jest.fn();
 
-    const paginationState = await createClient().iteratePreventionPolicyMemberIds(
-      {
+    const paginationState =
+      await createClient().iteratePreventionPolicyMemberIds({
         callback: cbSpy,
-        policyId: "40bb0ba06b9f4a10a4330fccecc01f84",
-      },
-    );
+        policyId: '40bb0ba06b9f4a10a4330fccecc01f84',
+      });
 
     expect(paginationState).toEqual({
       finished: true,
@@ -616,17 +689,20 @@ describe("iteratePreventionPolicyMemberIds", () => {
     );
   }, 20000);
 
-  test("partial set in multiple callbacks", async () => {
-    p = polly(__dirname, "iteratePreventionPolicyMemberIdsCompletes");
+  test('partial set in multiple callbacks', async () => {
+    recording = setupCrowdstrikeRecording({
+      directory: __dirname,
+      name: 'iteratePreventionPolicyMemberIdsCompletes',
+    });
+
     const cbSpy = jest.fn();
 
-    const paginationState = await createClient().iteratePreventionPolicyMemberIds(
-      {
+    const paginationState =
+      await createClient().iteratePreventionPolicyMemberIds({
         callback: cbSpy,
         pagination: { limit: 1 },
-        policyId: "40bb0ba06b9f4a10a4330fccecc01f84",
-      },
-    );
+        policyId: '40bb0ba06b9f4a10a4330fccecc01f84',
+      });
 
     expect(paginationState).toEqual({
       finished: true,