+++ src/steps/fetchDevicePolicyRelationships.ts
@@ -1,120 +1,77 @@
 import {
+  getRawData,
   IntegrationStepExecutionContext,
-  IntegrationStepIterationState,
-  RelationshipFromIntegration,
-} from "@jupiterone/jupiter-managed-integration-sdk";
-
+  RelationshipClass,
+  Step,
+} from "@jupiterone/integration-sdk-core";
+import { Entities, Relationships, StepIds } from "../constants";
 import createFalconAPIClient from "../crowdstrike/createFalconAPIClient";
-import { getPageLimit } from "../crowdstrike/pagination";
-import {
-  NumericOffsetPaginationParams,
-  NumericOffsetPaginationState,
-} from "../crowdstrike/types";
-import getIterationState from "../getIterationState";
-import { SENSOR_AGENT_PREVENTION_POLICY_RELATIONSHIP_TYPE } from "../jupiterone/converters";
-import ProviderGraphObjectCache from "../ProviderGraphObjectCache";
+import { PreventionPolicy } from "../crowdstrike/types";
+import { CrowdStrikeIntegrationInstanceConfig } from "../types";
 
-const MEMBERS_PAGINATION: NumericOffsetPaginationParams = {
-  limit: getPageLimit("policy-members", 100),
-};
+export async function fetchDevicePolicyRelationships(
+  context: IntegrationStepExecutionContext<
+    CrowdStrikeIntegrationInstanceConfig
+  >,
+): Promise<void> {
+  const { instance, jobState, logger } = context;
+  const client = createFalconAPIClient(instance.config, logger);
 
-export default {
-  id: "fetch-device-policies",
-  name: "Fetch Device Policies",
-  iterates: true,
-  executionHandler: async (
-    executionContext: IntegrationStepExecutionContext,
-  ): Promise<IntegrationStepIterationState> => {
-    const { logger } = executionContext;
+  logger.info("Iterating policy members...");
+  await jobState.iterateEntities(
+    { _type: Entities.PREVENTION_POLICY._type },
+    async preventionPolicyEntity => {
+      const preventionPolicy = getRawData<PreventionPolicy>(
+        preventionPolicyEntity,
+      );
+      if (!preventionPolicy) {
+        logger.warn(
+          {
+            _key: preventionPolicyEntity._key,
+          },
+          "Could not get prevention policy raw data from entity.",
+        );
+        return;
+      }
 
-    const cache = executionContext.clients.getCache();
-    const objectCache = new ProviderGraphObjectCache(cache);
-    const falconAPI = createFalconAPIClient(executionContext);
+      const policyId = preventionPolicy.id;
 
-    const iterationState = getIterationState(executionContext);
-
-    logger.info({ iterationState }, "Iterating policy members...");
-
-    const deviceIds = new Set<string>(
-      (await cache.getEntry("device-ids")).data || [],
-    );
-    const policyIds: string[] =
-      (await cache.getEntry("prevention-policy-ids")).data || [];
-
-    let policyPagination: NumericOffsetPaginationState = iterationState.state
-      .policyPagination || {
-      total: policyIds.length,
-      seen: 0,
-      offset: 0,
-      finished: false,
-    };
-
-    let policyIndex = policyPagination.offset || 0;
-
-    let membersPagination: NumericOffsetPaginationParams = iterationState.state
-      .membersPagination || { ...MEMBERS_PAGINATION };
-
-    do {
-      const policyId = policyIds[policyIndex];
-      const loggerInfo = { policyPagination, membersPagination };
-
-      membersPagination = await falconAPI.iteratePreventionPolicyMemberIds({
+      await client.iteratePreventionPolicyMemberIds({
+        // TODO remove `pagination` arg
+        pagination: undefined,
+        policyId: policyId,
         callback: async memberIds => {
-          logger.trace(loggerInfo, "Processing page of member ids");
-
-          const relationships: RelationshipFromIntegration[] = [];
+          logger.info(
+            {
+              memberCount: memberIds.length,
+              policyId,
+            },
+            "Processing page of member ids",
+          );
 
           for (const deviceId of memberIds) {
-            if (deviceIds.has(deviceId)) {
-              relationships.push({
-                _key: `${deviceId}|assigned|${policyId}`,
-                _type: SENSOR_AGENT_PREVENTION_POLICY_RELATIONSHIP_TYPE,
-                _class: "ASSIGNED",
-                _fromEntityKey: deviceId,
-                _toEntityKey: policyId,
-                displayName: "ASSIGNED",
-              });
-            }
+            await jobState.addRelationship({
+              _key: `${deviceId}|assigned|${policyId}`,
+              _type: Relationships.SENSOR_ASSIGNED_PREVENTION_POLICY._type,
+              _class: RelationshipClass.ASSIGNED,
+              _fromEntityKey: deviceId,
+              _toEntityKey: policyId,
+              displayName: "ASSIGNED",
+            });
           }
-          await objectCache.putRelationships(relationships);
         },
-        pagination: membersPagination,
-        policyId,
       });
+    },
+  );
+}
 
-      if (membersPagination.finished) {
-        membersPagination = MEMBERS_PAGINATION;
-
-        policyPagination = {
-          ...policyPagination,
-          offset: policyIndex,
-          seen: policyPagination.seen + 1,
-          finished: policyIndex === policyIds.length - 1,
-        };
-
-        logger.info(
-          { policyPagination, membersPagination },
-          "Finished paging policy members",
-        );
-
-        policyIndex += 1;
-      }
-    } while (!policyPagination.finished);
-
-    await objectCache.putCollectionStates({
-      type: SENSOR_AGENT_PREVENTION_POLICY_RELATIONSHIP_TYPE,
-      success: policyPagination.finished && !!membersPagination.finished,
-    });
-
-    return {
-      ...iterationState,
-      finished: policyPagination.finished,
-      state: {
-        policyPagination,
-        membersPagination: policyPagination.finished
-          ? undefined
-          : membersPagination,
-      },
-    };
-  },
+export const fetchDevicePolicyRelationshipsStep: Step<IntegrationStepExecutionContext<
+  CrowdStrikeIntegrationInstanceConfig
+>> = {
+  id: StepIds.DEVICE_POLICY_RELATIONSHIPS,
+  name: "Fetch Device Policies",
+  entities: [],
+  relationships: [Relationships.SENSOR_ASSIGNED_PREVENTION_POLICY],
+  dependsOn: [StepIds.DEVICES, StepIds.PREVENTION_POLICIES],
+  executionHandler: fetchDevicePolicyRelationships,
 };