+++ src/steps/fetchDevices.ts
@@ -1,96 +1,53 @@
 import {
-  createIntegrationRelationship,
-  EntityFromIntegration,
-  IntegrationRelationship,
+  createDirectRelationship,
   IntegrationStepExecutionContext,
-  IntegrationStepIterationState,
-} from "@jupiterone/jupiter-managed-integration-sdk";
-
+  RelationshipClass,
+  Step,
+} from "@jupiterone/integration-sdk-core";
+import { Entities, Relationships, StepIds } from "../constants";
 import createFalconAPIClient from "../crowdstrike/createFalconAPIClient";
-import getIterationState from "../getIterationState";
-import {
-  ACCOUNT_SENSOR_AGENT_RELATIONSHIP_TYPE,
-  createSensorAgentEntity,
-  SENSOR_AGENT_ENTITY_TYPE,
-} from "../jupiterone/converters";
-import ProviderGraphObjectCache from "../ProviderGraphObjectCache";
+import { createSensorAgentEntity } from "../jupiterone/converters";
+import { CrowdStrikeIntegrationInstanceConfig } from "../types";
+import { getAccountEntityFromJobState } from "./getAccount";
 
-export default {
-  id: "fetch-devices",
-  name: "Fetch Devices",
-  iterates: true,
-  executionHandler: async (
-    executionContext: IntegrationStepExecutionContext,
-  ): Promise<IntegrationStepIterationState> => {
-    const { logger } = executionContext;
+export async function fetchDevices(
+  context: IntegrationStepExecutionContext<
+    CrowdStrikeIntegrationInstanceConfig
+  >,
+): Promise<void> {
+  const { instance, jobState, logger } = context;
 
-    const cache = executionContext.clients.getCache();
-    const objectCache = new ProviderGraphObjectCache(cache);
+  const accountEntity = await getAccountEntityFromJobState(jobState);
+  const client = createFalconAPIClient(instance.config, logger);
 
-    const accountEntity = await objectCache.getAccount();
-    const deviceIds = (await cache.getEntry("device-ids")).data || [];
+  logger.info("Iterating devices...");
+  await client.iterateDevices({
+    // TODO remove `pagination` arg
+    pagination: undefined,
+    query: {
+      filter: `last_seen:>='${lastSeenSince()}'`,
+    },
+    callback: async devices => {
+      logger.info(
+        { deviceCount: devices.length },
+        "Creating device entities and relationships...",
+      );
 
-    const falconAPI = createFalconAPIClient(executionContext);
-
-    const iterationState = getIterationState(executionContext);
-
-    logger.info({ iterationState }, "Iterating devices...");
-
-    const filter =
-      iterationState.iteration === 0
-        ? `last_seen:>='${lastSeenSince()}'`
-        : undefined;
-
-    const pagination = await falconAPI.iterateDevices({
-      callback: async devices => {
-        logger.info(
-          { deviceCount: devices.length },
-          "Creating device entities and relationships...",
+      for (const device of devices) {
+        const deviceEntity = await jobState.addEntity(
+          createSensorAgentEntity(device),
         );
-
-        const sensorEntities: EntityFromIntegration[] = [];
-        const sensorRelationships: IntegrationRelationship[] = [];
-        for (const device of devices) {
-          const entity = createSensorAgentEntity(device);
-          sensorEntities.push(entity);
-          sensorRelationships.push(
-            createIntegrationRelationship({
-              _class: "HAS",
-              from: accountEntity,
-              to: entity,
-            }),
-          );
-        }
-        await Promise.all([
-          objectCache.putEntities(sensorEntities),
-          objectCache.putRelationships(sensorRelationships),
-        ]);
-      },
-      pagination: iterationState.state.pagination,
-      query: {
-        filter,
-      },
-    });
-
-    await cache.putEntry({ key: "device-ids", data: deviceIds });
-
-    await objectCache.putCollectionStates(
-      { type: SENSOR_AGENT_ENTITY_TYPE, success: pagination.finished },
-      {
-        type: ACCOUNT_SENSOR_AGENT_RELATIONSHIP_TYPE,
-        success: pagination.finished,
-      },
-    );
-
-    return {
-      ...iterationState,
-      finished: pagination.finished,
-      state: {
-        pagination,
-      },
-    };
-  },
-};
+        await jobState.addRelationship(
+          createDirectRelationship({
+            from: accountEntity,
+            _class: RelationshipClass.HAS,
+            to: deviceEntity,
+          }),
+        );
+      }
+    },
+  });
+}
 
 const THIRTY_DAYS_AGO = 60 * 1000 * 60 * 24 * 30;
 const LAST_SEEN_DAYS_BACK = THIRTY_DAYS_AGO;
@@ -98,3 +55,14 @@ const LAST_SEEN_DAYS_BACK = THIRTY_DAYS_AGO;
 function lastSeenSince(): string {
   return new Date(Date.now() - LAST_SEEN_DAYS_BACK).toISOString();
 }
+
+export const fetchDevicesStep: Step<IntegrationStepExecutionContext<
+  CrowdStrikeIntegrationInstanceConfig
+>> = {
+  id: StepIds.DEVICES,
+  name: "Fetch Devices",
+  entities: [Entities.SENSOR],
+  relationships: [Relationships.ACCOUNT_HAS_SENSOR],
+  dependsOn: [StepIds.ACCOUNT],
+  executionHandler: fetchDevices,
+};