+++ src/steps/fetchPreventionPolicies.ts
@@ -1,91 +1,61 @@
 import {
-  createIntegrationRelationship,
-  EntityFromIntegration,
-  IntegrationRelationship,
+  createDirectRelationship,
   IntegrationStepExecutionContext,
-  IntegrationStepIterationState,
-} from "@jupiterone/jupiter-managed-integration-sdk";
-
+  RelationshipClass,
+  Step,
+} from "@jupiterone/integration-sdk-core";
+import { Entities, Relationships, StepIds } from "../constants";
 import createFalconAPIClient from "../crowdstrike/createFalconAPIClient";
-import getIterationState from "../getIterationState";
-import {
-  createPreventionPolicyEntity,
-  PREVENTION_POLICY_ENFORCES_PROTECTION_RELATIONSHIP_TYPE,
-  PREVENTION_POLICY_ENTITY_TYPE,
-} from "../jupiterone/converters";
-import ProviderGraphObjectCache from "../ProviderGraphObjectCache";
+import { createPreventionPolicyEntity } from "../jupiterone/converters";
+import { CrowdStrikeIntegrationInstanceConfig } from "../types";
+import { getProtectionServiceEntityFromJobState } from "./getAccount";
 
-export default {
-  id: "fetch-prevention-policies",
-  name: "Fetch Prevention Policies",
-  iterates: true,
-  executionHandler: async (
-    executionContext: IntegrationStepExecutionContext,
-  ): Promise<IntegrationStepIterationState> => {
-    const { logger } = executionContext;
+export async function fetchPreventionPolicies(
+  context: IntegrationStepExecutionContext<
+    CrowdStrikeIntegrationInstanceConfig
+  >,
+): Promise<void> {
+  const { instance, jobState, logger } = context;
 
-    const cache = executionContext.clients.getCache();
-    const objectCache = new ProviderGraphObjectCache(cache);
+  const protectionServiceEntity = await getProtectionServiceEntityFromJobState(
+    jobState,
+  );
+  const client = createFalconAPIClient(instance.config, logger);
 
-    const protectionService = (
-      await cache.getEntry("endpoint-protection-service")
-    ).data;
+  logger.info("Iterating protection policies...");
+  await client.iteratePreventionPolicies({
+    // TODO remove `pagination` arg
+    pagination: undefined,
+    callback: async preventionPolicies => {
+      logger.info(
+        { policyCount: preventionPolicies.length },
+        "Creating protection policy entities and relationships...",
+      );
 
-    const policyIds =
-      (await cache.getEntry("prevention-policy-ids")).data || [];
-
-    const falconAPI = createFalconAPIClient(executionContext);
-
-    const iterationState = getIterationState(executionContext);
-
-    logger.info({ iterationState }, "Iterating protection policies...");
-
-    const pagination = await falconAPI.iteratePreventionPolicies({
-      callback: async policies => {
-        logger.info(
-          { policyCount: policies.length },
-          "Creating protection policy entities and relationships...",
+      for (const preventionPolicy of preventionPolicies) {
+        const preventionPolicyEntity = await jobState.addEntity(
+          createPreventionPolicyEntity(preventionPolicy),
         );
+        await jobState.addRelationship(
+          createDirectRelationship({
+            from: preventionPolicyEntity,
+            // TODO add ENFORCES to RelationshipClass
+            _class: "ENFORCES" as RelationshipClass,
+            to: protectionServiceEntity,
+          }),
+        );
+      }
+    },
+  });
+}
 
-        const entities: EntityFromIntegration[] = [];
-        const relationships: IntegrationRelationship[] = [];
-        for (const policy of policies) {
-          const entity = createPreventionPolicyEntity(policy);
-          entities.push(entity);
-          relationships.push(
-            createIntegrationRelationship({
-              _class: "ENFORCES",
-              from: entity,
-              to: protectionService,
-            }),
-          );
-          policyIds.push(policy.id);
-        }
-        await Promise.all([
-          await objectCache.putEntities(entities),
-          await objectCache.putRelationships(relationships),
-        ]);
-      },
-      pagination: iterationState.state.pagination,
-    });
-
-    await cache.putEntry({ key: "prevention-policy-ids", data: policyIds });
-
-    await objectCache.putCollectionStates(
-      {
-        type: PREVENTION_POLICY_ENTITY_TYPE,
-        success: pagination.finished,
-      },
-      {
-        type: PREVENTION_POLICY_ENFORCES_PROTECTION_RELATIONSHIP_TYPE,
-        success: pagination.finished,
-      },
-    );
-
-    return {
-      ...iterationState,
-      finished: pagination.finished,
-      state: { pagination },
-    };
-  },
+export const fetchPreventionPoliciesStep: Step<IntegrationStepExecutionContext<
+  CrowdStrikeIntegrationInstanceConfig
+>> = {
+  id: StepIds.PREVENTION_POLICIES,
+  name: "Fetch Prevention Policies",
+  entities: [Entities.PREVENTION_POLICY],
+  relationships: [Relationships.PREVENTION_POLICY_ENFORCES_PROTECTION_SERVICE],
+  dependsOn: [StepIds.ACCOUNT],
+  executionHandler: fetchPreventionPolicies,
 };