+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- Transitioned project from using `@jupiterone/jupiter-managed-integration-sdk`
+  to open-source `@jupiterone/integration-sdk-*` packages
+  
 ## 1.4.1 - 2021-08-26
 
 ### Fixed